requirements.txt
blessings==1.7
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
drawille==0.1.0
gpustat==0.6.0
huepy==1.2.1
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
npyscreen==4.10.5
nvidia-ml-py3==7.352.0
pandas==1.2.0
Pillow==8.0.1
psutil==5.7.3
ptflops @ git+https://github.com/sovrasov/flops-counter.pytorch.git@469b7430ec7c6aa8c258da1bca2c04de81fc9613
ptop==1.0
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
thop @ git+https://github.com/Lyken17/pytorch-OpCounter.git@d1920d3400a62678ab3c2cf21ba13db35921713c
timm==0.3.2
torch==1.7.1
torchstat==0.0.7
torchvision==0.8.2
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
yacs==0.1.8
